Reorganise the application icons
After looking at this in a bit of detail, the way that the icons were
just dumped in a single directory didn't make sense; well, not if you
want them to be easy to maintain over the course of time. So, this
change organises them in a more meaningful way.