Refactor the search form
This change fixes up the search form so that it is semantically correct,
such as using labels for all form elements, not attempting using HTML
elements where they are not allowed, and constructing the form in a more
responsive-friendly way. It also makes some minor changes for
accessibility, such as increasing the text size of the weather format
descriptions, increasing the contrast of all text in the form, and
removing positive tabindex values.