Basic refactor to make the site more mobile responsive
While not complete, this change starts the process of making the site
more responsive on mobile devices. It adjusts the width so that the
content fits within the view area of whatever device it is being
displayed on, and changes the layout approach so that it makes sense on
said device, such as using a column for mobile devices, but a row for
larger devices/screens.